Add two-pass union finding
The first pass creates a union of cross sections by placing
intersection-planes at regular intervals. This is not guaranteed to hit
all features of the input mesh. Thus a second pass is performed.
In this second pass additional cross sections are added to the union for
all vertices whose projection does not fall into the union that was
found so far.

This increases the runtime of the algorithm, but should guarantee accurate
results.