Add grid_sample method
The grid_sample methods chooses interesting points on a polyline
and then tries to "help" the mesher by treating these points as
intersections of an imagined grid and adding the remaining intersections
between the segments of the polyline and the grid to the polyline
itself.

This was an idea to produce more regular meshes, but first experiments
werent succesfull.