Improve min-gap algorithm
The thickness measurement via maximal inscribed spheres is now more
stable. It now considers the maximum discretization error (in the
maximum norm) when deciding wether to shrink a sphere further.
This produces good distances on allmost all meshes of the current test
set. The closest points are still sometimes wrong, especially when
determining thicknesses in large cylindrical segments.

The min-gap calculation is now score based, instead of punishing the
mesh diameter directly. The scoring considers
* topological distance
* gap size
* similarity of surface normals
* dihedral angle between triangles near the gap origin

The min-gap is then determined as the minimum gap of all gaps within a
certain score percentile. Considering the best 20 percent of all gaps
seems to produce good results across a range of meshes.