Improve precision of maximal inscribed spheres
Fixes the discretization error that is used by the MIS calculation to
determine whether a sphere should be shrunk further. Previously the
discretization error was only considered once. But the error occurs both
when determining the sphere center and when testing points. This means
opposite points on the sphere might deviate from the sphere radius by up
to two times the discretization error.

Also starts using vertex normals again. Together this leads to much
better results.