Refactor into a proper library
This commit:
- introduces a first version of a proper API in src/hexmesher.hpp
- moves all of the report features from the hexmesher application to the
  hexmesher library
- renames the hexmesher application to hexmesher-cli to avoid confusion
  with the library
- adds a "warnings" command to hexmesher-cli, which lists badly behaved
  triangles. Currently warns about self-intersections, degenerate
  triangles, and triangles with high aspect ratio.
- generally cleans up some code

The new api is a minmal API that for now supports determining min-gaps
and accessing the information required by the report and warning
commands. For now the API is fully CGAL independent. Any CGAL types are
hidden via the PIMPL idiom. The future will show if that continues to be
viable.