Disable installation if any public dependencies were fetched
Installing MeshHexer if any of the public dependencies were fetched
during the build is difficult. Because the dependencies are part of our
build, we are linking against internal targets. We thus would need to
install these targets alongside MeshHexer and ensure our link interface
refers to the installed libraries and not our internal targets.

Because that is just a huge bother, we have decided to only support
installation if all dependencies can be found on the system via normal
find_package calls.

Private dependencies, i.e. dependencies of targets that are not in any
export set, such as Catch2 do not cause these problems.

We now differentiate between public and private dependencies in get_tpl
and report any fetched public dependencies to the parent scope.