Several changes in build system and installation
This commit introduce several changes to the build system and
more significantly to the install behavior.

First of all, the public header are now in a seperate include folder
and including them now requires #include <meshhexer/...>.

Also, we removed the pmp dependency for now.

Furtermore, installing our target is now consistent regarding thirdparty
dependencies. To accomplish that, installing now requires
all PUBLIC thirdparty dependencies, i.e. dependencies which
are linked to the MeshHexer library, to be findable by a
find_package call, which means, you have to install all
thirdparty dependencies yourself before installing MeshHexer and
provide a way for CMake to find those, e.g. set
CMAKE_PREFIX_PATH accordingly.

This is necessary, since exporting all thirdparty dependencies
(especially boost) is a herculean task, which breaks as soon as
a user changes dependencies on his end.

This is also more consistent to the way a package manager would expect
the dependencies to be handled.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>