Improve FBM mesh generator
Replaces the previous complex algorithms with an octree based approach.
Based on the intended number of refinements of the mesh hierarchy, a
field of target cell sizes for the base mesh is computed. That field is
then inserted into an octree, such that the octrees cells are split as
long as they are larger then the target cell size of the sample points
they contain.

The deepest complete layer of the octree then determines the number of
global refinements. All other leaves are used to determine local
refinement levels, so that target sizes are hit across the final base
mesh.

Also adds the ability to set a custom bounding box for the base mesh,
and the filenames for the output files.